+++ changelog.md
@@ -3,3 +3,4 @@
 16th february 2026:
 
 - started electron app!
+- removed bar at the top and prepared index.html file